doing some testing to see how to parse out all specials of all shows, ideally into SQL
